refactor: user config
