+++ crates/pm/src/cmd/install.rs
@@ -2,8 +2,8 @@ use anyhow::Result;
 use std::path::Path;
 
 use crate::service::install::InstallService;
-use crate::util::settings::get_omit;
 use crate::util::save_type::{PackageAction, SaveType};
+use crate::util::user_config::get_omit;
 
 pub async fn update_packages(
     action: PackageAction,
