+++ crates/pm/src/cmd/ping.rs
@@ -1,12 +1,11 @@
 use anyhow::Result;
 use colored::Colorize;
 
-use crate::util::settings::{detect_supports_semver, get_registry};
+use crate::util::registry::ping_registry;
+use crate::util::user_config::{detect_supports_semver, get_registry};
 
 pub async fn ping(registry: Option<&str>) -> Result<()> {
-    let registry = registry
-        .map(String::from)
-        .unwrap_or_else(get_registry);
+    let registry = registry.map(String::from).unwrap_or_else(get_registry);
 
     println!("{} {}", "PING".green(), registry.cyan());
 
@@ -14,38 +13,27 @@ pub async fn ping(registry: Option<&str>) -> Result<()> {
         .timeout(std::time::Duration::from_secs(10))
         .build()?;
 
-    let ping_url = format!("{}/-/ping", registry);
-    let start = std::time::Instant::now();
+    let result = ping_registry(&client, &registry).await;
 
-    match client.get(&ping_url).send().await {
-        Ok(resp) => {
-            let latency = start.elapsed().as_millis();
-            if resp.status().is_success() {
-                let supports = detect_supports_semver(&registry).await;
-                let semver_info = if supports {
-                    "supports-semver: yes".green()
-                } else {
-                    "supports-semver: no".yellow()
-                };
-                println!(
-                    "{} {}ms ({})",
-                    "PONG".green(),
-                    latency.to_string().cyan(),
-                    semver_info
-                );
-            } else {
-                println!(
-                    "{} HTTP {} ({}ms)",
-                    "FAIL".red(),
-                    resp.status(),
-                    latency
-                );
-            }
-        }
-        Err(e) => {
-            let latency = start.elapsed().as_millis();
-            println!("{} {} ({}ms)", "FAIL".red(), e, latency);
-        }
+    if result.success {
+        let supports = detect_supports_semver(&registry).await;
+        let semver_info = if supports {
+            "supports-semver: yes".green()
+        } else {
+            "supports-semver: no".yellow()
+        };
+        println!(
+            "{} {}ms ({})",
+            "PONG".green(),
+            result.latency_ms.to_string().cyan(),
+            semver_info
+        );
+    } else {
+        println!(
+            "{} registry did not respond ({}ms)",
+            "FAIL".red(),
+            result.latency_ms
+        );
     }
 
     Ok(())
