+++ crates/pm/src/util/downloader.rs
@@ -10,10 +10,10 @@ use tokio::sync::Semaphore;
 use tokio_retry::RetryIf;
 
 use super::cache::get_cache_dir;
-use super::settings::get_manifests_concurrency_limit_sync;
 use super::extractor::extract_and_write;
 use super::oncemap::OnceMap;
 use super::retry::{RetryableError, build_dns_cached_client, create_retry_strategy};
+use super::user_config::get_manifests_concurrency_limit_sync;
 
 // Global downloader client - no pool limit, concurrency controlled by OnceMap
 static DOWNLOADER_CLIENT: Lazy<Client> = Lazy::new(build_dns_cached_client);
